Refactor Ensembl-specific code into separate module
